+++ ovenPlatform-be/server.js
@@ -8,6 +8,7 @@ import streamRoutes from "./routes/streamRoutes.js";
 import authRoutes from "./routes/authRoutes.js";
 import userAuthRoutes from "./routes/userAuthRoutes.js";
 import profileRoutes from "./routes/profileRoutes.js";
+import monitoringRoutes from "./routes/monitoringRoutes.js";
 
 const app = express();
 const PORT = config.NodePort;
@@ -45,6 +46,7 @@ app.use("/v1/ome/streams", streamRoutes);
 app.use("/v1/ome", authRoutes); // Routes for OvenMediaEngine connection
 app.use("/v1/auth", userAuthRoutes); // Routes for user authentication
 app.use("/v1/profile", profileRoutes); // Routes for user profile management
+app.use("/v1/monitoring", monitoringRoutes); // Routes for stream monitoring management
 
 // Debug endpoint
 app.get("/v1/status", (req, res) => {